# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookings Strapblaque API Documentation",
        "description": "This is the API documentation for the application.",
        "contact": {
            "email": "info@strapblaque.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/subscription/plans": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "List all active subscription plans",
                "description": "Returns a list of all active subscription plans.",
                "operationId": "c86411aea0da8a4011ca7f20a1d1dc54",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/subscriptions": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Get current subscription details",
                "description": "Returns the current subscription details for the authenticated organization.",
                "operationId": "8473d089bae310cee331da8d711850c1",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Subscribe to a new plan",
                "description": "Subscribes the authenticated organization to a new plan.",
                "operationId": "a71648bdc4f0b03d53530ab22adf00cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plan_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/subscriptions/cancel": {
            "put": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Cancel the current subscription",
                "description": "Cancels the current subscription for the authenticated organization.",
                "operationId": "e56244043c3b917c19715933f2dec354",
                "responses": {
                    "200": {
                        "description": "Subscription canceled successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/subscriptions/initiate-payment": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Initiate payment for a subscription plan",
                "description": "Initiates payment for a subscription plan and redirects to the payment gateway.",
                "operationId": "931509bb1c90add2dca4f74b60c37bb4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plan_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect to payment gateway"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/subscriptions/payment-notify": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Handle payment notifications",
                "description": "Handles payment notifications from the payment gateway.",
                "operationId": "b345d738cf9ffdb4eafdbf8d3a226741",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment_status": {
                                        "type": "string",
                                        "example": "COMPLETE"
                                    },
                                    "custom_str1": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "custom_str2": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "signature": {
                                        "type": "string",
                                        "example": "abc123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment notification processed"
                    },
                    "400": {
                        "description": "Invalid signature"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/subscriptions/success": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Handle payment success",
                "description": "Handles successful payment and returns a success message.",
                "operationId": "03f5e1a30bc2074e0f3ff1090476865a",
                "responses": {
                    "200": {
                        "description": "Subscription activated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/subscriptions/cancel": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Handle payment cancellation",
                "description": "Handles payment cancellation and returns a cancellation message.",
                "operationId": "cbcde7f2c1a2b05511986998a9bae021",
                "responses": {
                    "200": {
                        "description": "Payment was canceled"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
...
}